Extract EPUB cover menu parts
